requirements.txt
Django==5.0
google-generativeai
Pillow
PyPDF2
